+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # CHANGELOG
 
 ## 2.0.0
-## BREAKING
+### BREAKING
 * Wrap `Aws::S3::Errors::NotFound` from `Attachie::S3Driver` and
  `Errno::ENOENT` from `Attachie::FileDriver` in `Attachie::ItemNotFound`
 * Change `Attachie::FakeDriver::ItemNotFound` to `Attachie::ItemNotFound`
 * Make `#delete` always return `true` for all drivers
+### Added
+* Added `Attachie::BaseError` as a base for all attachie exceptions
 
 ## 1.2.1
 ### Fixed